Move json client to dvotecli
